scanning with VT api now fixed (I hope-) v3.1
