+++ settings/s.ini
@@ -2,8 +2,8 @@
 default_theme = dark_cyan.xml
 
 [Config-Settings]
-use_vt_api = False
+use_vt_api = True
 use_mt_api = False
-vt_api_key = 
+vt_api_key = b35ef7b73feb8dc88e519f4a610af4cedd015765cddaef3426521e0b92b14966
 mt_api_key =
